move reverse row into reverse component
